+++ changelog.md
@@ -6,6 +6,7 @@ __1.1.2__
 * inline, at compile-time, more template functions (e.g `concat`) for faster execution
 * inlined plugins can receive passed arguments (via `$` array parameter) both as (comma-separated) string (`$0`), plus each separate argument (`$1..$n`)
 * lazy init and rendering for inline templates and plugins for even faster execution
+* remove `plural` functions and add liguisticaly contexual localisation `xlocale`, `nxlocale` functions
 * `cc` function alias of `concat`
 * fix php-like `date()` function in python implementation (wrong day of week for `Sunday`)